Add condition to only nominate restore on DT build success.
If there are changes in a failing DT build, remember to unconditionally nominate restore the next time DT build succeeds. This unconditional nomination may be redundant if nothing actually changed (i.e. the change we saw in the failing DT build was spurious), but it will always be correct -- and NuGet can handle redundant nominations.

One problem with this solution: if restore adds some targets to *.nuget.g.targets file, and those targets fail DT build, then we won't get another restore until that failure is resolved. If the resolution is "remove the package that added the offending target", then we have a chicken and egg problem: restore needs to run to delete offending target from *.nuget.g.targets, but target needs to be deleted to pass DT build and free restore to run. Meanwhile no other project changes will trigger restore until the DT build failure is resolved.

A better solution would be to condition restore nomination on the success of the subset of targets we care about (instead of entire DT build). However this chicken and egg problem would still exist if the offending target above prevents a target from this subset from running.